Several changes made to dbDefs.h
The definitions for the macros YES, NO, NONE, min() and max() have been deleted. YES and NO have been replaced by menuYesNoYES and menuYesNoNO from the menuYesNo.h file where they were used in several record types. The other macros were not being used anywhere in Base, sncseq or Asyn.

The macro LOCAL that was a synonym for static is now deprecated and will be deleted in R3.15, please adjust your code to use the latter keyword. All uses of the READONLY macro from shareLib.h must now be replaced by the keyword const as the macro has been deleted.

The dbDefs.h file was unnecessarily including various other libCom header files which may have to be manually added to out-of-tree source files that relied on this. The general rule for header files is that a header should only include other headers that are needed for its own inclusion in any source file. The #include statements that might need to be added are:

    * #include <stdarg.h>
    * #include "errlog.h"
    * #include "errMdef.h"
    * #include "ellLib.h"
    * #include "epicsTypes.h"

A new macro CONTAINER(pointer, type, member) has been added which calculates and returns a pointer to the parent structure when given a pointer to a member, the structure type and the name of the member in that structure. On GNU compilers the type of the pointer is checked to ensure that it matches the member, so code using this macro should be built using gcc for additional confidence.